shop menu functions (visuals to be updated)
-added shop menu
-added buttons for purchases
-added dynamic price labels
-fixed upgrade manager errors
-changed prices to round to the nearest int
-changed resource ui to round to the nearest int, but not the actualy resource variable
-will update visuals when art is uploaded
-will continue to balance prices and effects

+++ Assets/Ashley/Scripts/ResourceManager.cs
@@ -13,7 +13,7 @@ public class ResourceManager : MonoBehaviour
     void Start()
     {
         meepleManager = gameObject.GetComponent<MeepleManager>();
-        resourceText.text = resources.ToString();
+        resourceText.text = ((int)resources).ToString();
         StartCoroutine(ResourceProduction());
     }
